chore: Move shared_types to utils/sharedtypes
