+++ .golangci.yaml
@@ -122,6 +122,8 @@ linters-settings:
         alias: clienttesting
       - pkg: github.com/kyma-project/telemetry-manager/internal/utils/logpipeline
         alias: logpipelineutils
+      - pkg: github.com/kyma-project/telemetry-manager/internal/utils/sharedtypes
+        alias: sharedtypesutils
   errcheck:
     check-type-assertions: true # Reports type assertions: `a := b.(SomeStruct)`.
     check-blank: true # Report assignment of errors to blank identifier: `num, _ := strconv.Atoi(numStr)`.
