+++ internal/fluentbit/config/builder/output.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
-	"github.com/kyma-project/telemetry-manager/internal/utils"
 	logpipelineutils "github.com/kyma-project/telemetry-manager/internal/utils/logpipeline"
+	sharedtypesutils "github.com/kyma-project/telemetry-manager/internal/utils/sharedtypes"
 )
 
 // Considering Fluent Bit's exponential back-off and jitter algorithm with the default scheduler.base and scheduler.cap,
@@ -60,17 +60,17 @@ func generateHTTPOutput(httpOutput *telemetryv1alpha1.LogPipelineHTTPOutput, fsB
 	sb.AddIfNotEmptyOrDefault("port", httpOutput.Port, "443")
 	sb.AddIfNotEmptyOrDefault("format", httpOutput.Format, "json")
 
-	if utils.IsValid(&httpOutput.Host) {
+	if sharedtypesutils.IsValid(&httpOutput.Host) {
 		value := resolveValue(httpOutput.Host, name)
 		sb.AddConfigParam("host", value)
 	}
 
-	if utils.IsValid(&httpOutput.Password) {
+	if sharedtypesutils.IsValid(&httpOutput.Password) {
 		value := resolveValue(httpOutput.Password, name)
 		sb.AddConfigParam("http_passwd", value)
 	}
 
-	if utils.IsValid(&httpOutput.User) {
+	if sharedtypesutils.IsValid(&httpOutput.User) {
 		value := resolveValue(httpOutput.User, name)
 		sb.AddConfigParam("http_user", value)
 	}
@@ -89,15 +89,15 @@ func generateHTTPOutput(httpOutput *telemetryv1alpha1.LogPipelineHTTPOutput, fsB
 
 	sb.AddConfigParam("tls.verify", tlsVerify)
 
-	if utils.IsValid(httpOutput.TLS.CA) {
+	if sharedtypesutils.IsValid(httpOutput.TLS.CA) {
 		sb.AddConfigParam("tls.ca_file", fmt.Sprintf("/fluent-bit/etc/output-tls-config/%s-ca.crt", name))
 	}
 
-	if utils.IsValid(httpOutput.TLS.Cert) {
+	if sharedtypesutils.IsValid(httpOutput.TLS.Cert) {
 		sb.AddConfigParam("tls.crt_file", fmt.Sprintf("/fluent-bit/etc/output-tls-config/%s-cert.crt", name))
 	}
 
-	if utils.IsValid(httpOutput.TLS.Key) {
+	if sharedtypesutils.IsValid(httpOutput.TLS.Key) {
 		sb.AddConfigParam("tls.key_file", fmt.Sprintf("/fluent-bit/etc/output-tls-config/%s-key.key", name))
 	}
 
@@ -109,7 +109,7 @@ func resolveValue(value telemetryv1alpha1.ValueType, logPipeline string) string 
 		return value.Value
 	}
 
-	if value.ValueFrom != nil && utils.IsValid(&value) {
+	if value.ValueFrom != nil && sharedtypesutils.IsValid(&value) {
 		secretKeyRef := value.ValueFrom.SecretKeyRef
 		return fmt.Sprintf("${%s}", FormatEnvVarName(logPipeline, secretKeyRef.Namespace, secretKeyRef.Name, secretKeyRef.Key))
 	}
