+++ internal/utils/sharedtypes/sharedtypes.go
@@ -1,4 +1,4 @@
-package utils
+package sharedtypes
 
 import (
 	telemetryv1alpha1 "github.com/kyma-project/telemetry-manager/apis/telemetry/v1alpha1"
